requirements.txt
click==8.1.7
colorama==0.4.6
pycryptodome==3.20.0
